pyproject.toml
[project]
name = "mail-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mcp[cli]>=1.9.4",
    "pytest>=8.4.1",
    "redis>=6.2.0",
    "setuptools>=80.9.0",
    "uvicorn>=0.34.3",
]
